pyproject.toml
[project]
name = "illico"
version = "0.1.0"
description = "Fast asymptotic mannwhitney-u test"
authors = [
    {name = "remydubois",email = "remydubois14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "numpy (>=2.3.4,<3.0.0)",
    "numba (>=0.62.1,<0.63.0)",
    "anndata (>=0.12.3,<0.13.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scipy (>=1.16.2,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "joblib (>=1.5.2,<2.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-benchmark (>=5.1.0,<6.0.0)",
    "pytest-benchmark[histogram] (>=5.2.3,<6.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
    "isort (>=7.0.0,<8.0.0)",
    "line-profiler (>=5.0.0,<6.0.0)",
    "pdex (>=0.1.26,<0.2.0)",
    "ipdb (>=0.13.13,<0.14.0)",
    "scanpy (>=1.11.5,<2.0.0)",
    "pytest-memray (>=1.8.0,<2.0.0)",
    "memory-profiler (>=0.61.0,<0.62.0)",
    "pytest-regex (>=0.2.0,<0.3.0)",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.ruff]
extend-exclude = ["tests", "notebooks"]

[tool.pytest.ini_options]
markers = [
    "debug: benchmarks on 1% of datasets for debugging",
    "speed_bench: slow speed benchmarks tests",
    "memory_bench: slow memory benchmarks tests",
]
addopts = "-p no:warnings"
